+++ core/execution/engine.go
@@ -1140,9 +1140,20 @@ func (e *Engine) OnTick(ctx context.Context, t time.Time) {
 		}
 		e.removeMarket(id)
 	}
+
+	// sort the marketCPStates by ID since the order we clear insurance pools
+	// changes the division when we split it across remaining markets and
+	// who the remainder ends up with if it doesn't divide equally.
+	allIDs := []string{}
+	for id := range e.marketCPStates {
+		allIDs = append(allIDs, id)
+	}
+	sort.Strings(allIDs)
+
 	// find state that should expire
-	for id, cpm := range e.marketCPStates {
+	for _, id := range allIDs {
 		// market field will be nil if the market is still current (ie not closed/settled)
+		cpm := e.marketCPStates[id]
 		if !cpm.TTL.Before(t) {
 			// CP data has not expired yet
 			continue